+++ GenerateSubsets.py
@@ -31,7 +31,7 @@ def binary_entropy(val):
     return -(val) * math.log2(val) - (1 - val) * (math.log2(1-val))
 
 def read_complex_conf(filepath):
-        bad_list = [28, 200, 503, 754]
+        bad_list = []
         with open(filepath, 'r') as f:
             confidence = []
             lines = f.readlines()
@@ -48,8 +48,15 @@ def read_complex_conf(filepath):
                     confidence.append(pair)
             return confidence, bad_list
 
+def gen(template,positions):
+    ret_value = []
+    for x in range(positions.shape[0]):
+        v_i = template[positions[x]]
+        ret_value.append(v_i)
+    return ret_value
 
 def sample_alpha(size, biometric_len, number_samples, confidence, alpha_param):
+    bad_list = [28, 200, 503, 754]
     if confidence is None:
         print("Can't run Smart sampling without confidence, calling uniform")
         return sample_uniform(size, biometric_len, number_samples, confidence)
@@ -75,6 +82,7 @@ def sample_alpha(size, biometric_len, number_samples, confidence, alpha_param):
     return np.array(sample_array)
 
 def sample_alpha_with_entropy(size, biometric_len, number_samples, confidence, alpha_param):
+    bad_list = [28, 200, 754]
     if confidence is None:
         print("Can't run Smart sampling without confidence, calling uniform")
         return sample_uniform(size, biometric_len, number_samples, confidence)
@@ -84,11 +92,12 @@ def sample_alpha_with_entropy(size, biometric_len, number_samples, confidence, a
 
     for set_selection_iter in range(number_samples):
         sample_indices = random.choices(range(len(new_confidence)), weights=new_confidence, k=size)
+        sample_indices = [index for index in sample_indices if index not in bad_list]
         dedup_indices = list(set(sample_indices))
         loop_count = 1
         while len(dedup_indices) < size:
             new_index = random.choices(range(len(new_confidence)), weights=new_confidence, k=max(1,size - len(dedup_indices)))
-            [dedup_indices.append(n) for n in new_index if n not in dedup_indices]
+            [dedup_indices.append(n) for n in new_index if n not in dedup_indices and n not in bad_list]
             loop_count = loop_count +1
             if loop_count == 1000000:
                 print("Smart sampling failed to find a non-duplicating subset")
@@ -101,23 +110,23 @@ def sample_alpha_with_entropy(size, biometric_len, number_samples, confidence, a
 ################################################################################
 
 # Command Line Usage:
-# python3 GenerateSubsetsNewFE.py [subset size] ['simple' or 'complex'] [alpha] [number of subsets] [output file name]
+# python3 GenerateSubsets.py [subset size] ['simple' or 'complex'] [alpha] [number of subsets] [output file name]
 
 size_or_threshold = int(sys.argv[1]) # Subset size
 selection_method = sys.argv[2] # 'complex' or 'simple'
-alpha_param = int(sys.argv[3]) # Confidence Weight Parameter
+alpha_param = float(sys.argv[3]) # Confidence Weight Parameter
 num_lockers = int(sys.argv[4]) # number of subsets sampled
-outputfilename = sys.argv[5] + str(size_or_threshold) +  str(selection_method) + str(alpha_param) + str(num_lockers) #output file name
+outputfilename = sys.argv[5] + str(size_or_threshold) +  str(selection_method) + str(alpha_param)  + '_' + str(num_lockers) #output file name
 numbers = re.compile(r'(\d+)')
 cwd = os.getcwd()
 num_cpus = mp.cpu_count()
-folder_list = sorted(glob.glob(cwd + "/CompFE/iris_right_all_dense_0.205_57-75_1024_folders/*"),key=numericalSort)
+folder_list = sorted(glob.glob(cwd + "<Enter your feature vector folder here>"),key=numericalSort)
 print (cwd)
 print ("Folders: ",len(folder_list))
 num_classes = range(len(folder_list))
 
 print("Reading Confidence")
-confidence, bad_list = read_complex_conf(cwd + "/CompFE/PythonImpl/AuxiliaryFiles/ConfidenceInfo.txt")
+confidence, bad_list = read_complex_conf(cwd + "/CompFE/PythonImpl/AuxiliaryFiles/confidence_672_70classes.txt")
 
 if selection_method == 'complex':
     print("Generating Subsets Using Complex Alpha Sampling")
