+++ README.md
@@ -16,3 +16,6 @@ Files:
 * PythonImpl/RPISetDifference.py implementation of set difference construction from https://eprint.iacr.org/2016/1100
 * Statistics folder is used to understand quality of zeta sampling algorithm.
 * subsets contains selected subsets for mid and high security parameters.
+* `GenerateSubsets.py` is a file that generates selected subsets and can be used as follows: `python3 GenerateSubsetsNewFE.py [subset size] ['simple' or 'complex'] [alpha] [number of subsets] [output file name]` and will be looking for a feature vector folder (which should be provided by the user) with each class having a folder and each vector having a file with a distinct filename that begins with the class name and a confidence file where each index has a line documenting the agreement with inter and intra class comparisons. Optionally, you may manually exclude indices from the complex confidence selection by providing indices in the bad_list variable. (Modification to `read_complex_conf` will be necessary to match your file structure and folder structure.)
+* `TarTest.py` is the test that calculates the entropy of a sample of your subsets and should be used as follows" `python3 EntropyTest.py [subset size or entropy threshold] [number of subsets] [subsets filename]` which takes as input the selected subsets from `GenerateSubsets.py`. See above for modifications.
+* `EntropyTest.py` is the test that calculates the TAR for your subsets and should be used as follows: `python3 EntropyTest.py [subset size] [number of subsets] [subsets filename] [subsets to test] [starting index]`. See above for modifications. Note this can take a very long time even parallelized which is why we offer a starting index.
